Fixed fucked relationships part 1 of 7

+++ db/migrate/20190614184004_create_users.rb
@@ -6,7 +6,7 @@ class CreateUsers < ActiveRecord::Migration[5.2]
       t.string :password_digest
       t.string :email
       t.integer :phone
-      t.integer :wallet , :default => 0
+      t.float :wallet , :default => 0.00
 
       t.timestamps
     end
